build: adding icon to the GUI window
Committing the resources for both an icon file, as well as a signature image. Adding this icon to the window within the ui file. New action has been added to the menu for log testing, this is intended as a temporary addition. Also, chore of removing the now outdated gitkeep.