feat: added filehandler to log, as well as one-line statusbar
Updates to the main from file to initialize and facilitate logging from two handlers. status_logger has been updated to set a single line of text, as opposed to scrolling - for the time being. An actions file has been created to handle the connected actions of controls on the GUI.