fix: Update gitignore to allow the gitkeep and log directory to remain
Previous gitignore configuration ignored all files within that subfolder, including the gitkeep. This resulted in the folder not being committed and causing some knock-on issues with the log file handler not finding the directory. Resolves the most recent commit. Also commits the gitkeep in question to allow for tracking of the logs folder.